UPDATE: Merged with current SVN source tree (2019-05-08)

+++ test/fakewap.c
@@ -1,7 +1,7 @@
 /* ==================================================================== 
  * The Kannel Software License, Version 1.0 
  * 
- * Copyright (c) 2001-2016 Kannel Group  
+ * Copyright (c) 2001-2019 Kannel Group
  * Copyright (c) 1998-2001 WapIT Ltd.   
  * All rights reserved. 
  * 
@@ -676,7 +676,7 @@ int send_post(int fd, unsigned short tid, int tid_new, char* url)
         panic(0, "fakewap: input file is empty");
     if (((transaction_mode==TXN_MODE_CONNECTION_LESS) && (fstats.st_size > (int)sizeof(msgbuf))) ||
         ((transaction_mode==TXN_MODE_CONNECTION_ORIENTED) && (fstats.st_size > 256*SAR_SEGM_SIZE))) {
-        panic(0, "fakewap: input file size (%ld) is too large", fstats.st_size);
+        panic(0, "fakewap: input file size (%ld) is too large", (long) fstats.st_size);
     }
             
     int nsegs = (fstats.st_size-1) / SAR_SEGM_SIZE;
@@ -741,7 +741,7 @@ int send_post(int fd, unsigned short tid, int tid_new, char* url)
         panic(0, "fakewap: input file read error, errno %d", errno);
     
     debug("fakewap", 0, "Sending WSP_POST, url %s, Content-Type %s, User-Agent %s, Content-Length %lu",
-          url, content_type, octstr_get_cstr(useragent) + 1, fstats.st_size );
+          url, content_type, octstr_get_cstr(useragent) + 1, (long unsigned) fstats.st_size );
     
     if (transaction_mode==TXN_MODE_CONNECTION_LESS) {
         ret = wap_msg_send( fd, NULL, 0, tid, tid_new,
